arrumando tela cadastro e adicionando middleware
