+++ database/register.json
@@ -1,99 +1,1 @@
-[
-  {
-    "primeironome": "Lenisa",
-    "sobrenome": "Mendes dos Santos",
-    "telefone": "13444444",
-    "email": "jcarlosmendes@outlook.com",
-    "cpf": "99999999999",
-    "aniversario": "2022-10-18",
-    "genero": "on",
-    "cep": "16201158",
-    "numero": "302",
-    "rua": "jose sanches garcia",
-    "cidade": "Birigüi",
-    "estado": "São Paulo",
-    "senha": "Birigui@17",
-    "confirmarSenha": "Birigui@17"
-  },
-  {
-    "primeironome": "Lenisa",
-    "sobrenome": "Mendes dos Santos",
-    "telefone": "13444444",
-    "email": "jcarlosmendes@outlook.com",
-    "cpf": "99999999999",
-    "aniversario": "2022-10-18",
-    "genero": "on",
-    "cep": "16201158",
-    "numero": "302",
-    "rua": "jose sanches garcia",
-    "cidade": "Birigüi",
-    "estado": "São Paulo",
-    "senha": "Birigui@17",
-    "confirmarSenha": "Birigui@17"
-  },
-  {
-    "primeironome": "Jose Carlos",
-    "sobrenome": "Mendes dos Santos",
-    "telefone": "7777777",
-    "email": "jcarlosmendes@outlook.kcom",
-    "cpf": "99999999999",
-    "aniversario": "2022-09-29",
-    "genero": "on",
-    "cep": "16201158",
-    "numero": "302",
-    "rua": "DR. carlos carvalho rosa",
-    "cidade": "Birigüi",
-    "estado": "São Paulo",
-    "senha": "Birigui@19",
-    "confirmarSenha": "Birigui@19"
-  },
-  {
-    "primeironome": "Jose Carlos",
-    "sobrenome": "Mendes dos Santos",
-    "telefone": "7777777",
-    "email": "jcarlosmendes@outlook.kcom",
-    "cpf": "99999999999",
-    "aniversario": "2022-10-16",
-    "genero": "on",
-    "cep": "16201158",
-    "numero": "302",
-    "rua": "jose sanches garcia",
-    "cidade": "Birigüi",
-    "estado": "São Paulo",
-    "senha": "Birigui@18",
-    "confirmarSenha": "Birigui@18"
-  },
-  {
-    "primeironome": "Lenisa",
-    "sobrenome": "Benti",
-    "telefone": "2222222",
-    "email": "jcarlosmendes@outlook.kcom",
-    "cpf": "99999999999",
-    "aniversario": "2022-10-03",
-    "genero": "on",
-    "cep": "55599999",
-    "numero": "1050",
-    "rua": "DR. carlos carvalho rosa",
-    "cidade": "Birigüi",
-    "estado": "São Paulo",
-    "senha": "Birigui@1",
-    "confirmarSenha": "Birigui@1"
-  },
-  { "email": "" },
-  { "email": "" },
-  {
-    "primeironome": "Jose Carlos",
-    "sobrenome": "Mendes dos Santos",
-    "telefone": "13444444",
-    "email": "jcarlosmendes@outlook.com",
-    "cpf": "444444444444",
-    "aniversario": "2022-11-23",
-    "genero": "on",
-    "cep": "16201158",
-    "numero": "302",
-    "rua": "jose sanches garcia",
-    "cidade": "Birigüi",
-    "estado": "São Paulo",
-    "senha": "Dh@123456"
-  }
-]
+
